fourthTask(add tests)

+++ src/com/ontopoff/listDir/FileOutputClass.java
@@ -1,5 +1,6 @@
 package com.ontopoff.listDir;
 
+import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -10,6 +11,10 @@ import java.nio.file.StandardOpenOption;
 public class FileOutputClass {
 
     public void checkFile(String path) throws IOException {
+        File outputFile = new File(path).getCanonicalFile();
+        if (!outputFile.exists()) {
+            throw new FileNotFoundException("Невозможно открыть файл вывода!");
+        }
         try(FileOutputStream fout = new FileOutputStream(path)) {
         } catch (FileNotFoundException e) {
             throw new FileNotFoundException("Невозможно открыть файл вывода!");
